add conversation wrapper
